PtAPI/requirements.txt
# Web framework
fastapi>=0.104.0,<0.116.0
uvicorn[standard]>=0.24.0,<0.32.0

# Database
sqlalchemy>=2.0.0
sqlalchemy-utils>=0.41.0
# Driver PostgreSQL (compatível com postgresql+psycopg:// na DATABASE_URL)
psycopg[binary]>=3.1.0

# Migrations
alembic>=1.12.0

# Config
python-dotenv>=1.0.0

# Auth & Security
passlib[argon2]>=1.7.4
python-jose[cryptography]>=3.3.0

# Validation
pydantic[email]>=2.5.0

# Rate limiting
slowapi>=0.1.9

# PDF generation
reportlab>=4.0.0
